refactor(skill-eval): consolidate entry point and simplify call sites
Fold eval.ts into index.ts, move getTags return type to Set<string>,
and move upsertComment try/catch into the function itself.
